+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.6.1-SNAPSHOT]
 
+- UIComponent.headingProperty is ObservableValue<String> for easier binding
 - `field` builder supports `orientation` parameter which will cause input container to be a VBox instead of an HBox (https://github.com/edvin/tornadofx/issues/190)
 - UIComponents can now be instantiated manually instead of via inject() and find()
 - Input Control builders now support ObservableValue instead of just Property for automatic binding